requirements.txt
scipy==1.6.3
pandas==1.2.4
matplotlib==3.4.1
numpy==1.19.5
mtcnn==0.1.0
Pillow==8.2.0
scikit_learn==0.24.2
tensorflow_gpu==2.4.0